style(package-manager): move the verifier test into an external module
The dylint perfectionist rule rejects an inline `mod tests { ... }`.
